DPR-81 update commentary rename python scripts
