+++ src/dev/_read_delta_table.py
@@ -1,11 +1,9 @@
-from time import strptime
+import boto3
 
-from pyspark.shell import spark
 from pyspark.sql.functions import *
 from pyspark.sql.types import DateType, TimestampType, Row
 from pyspark.sql import Window
 import datetime
-
 
 from pyspark.sql.types import (
     StringType,
@@ -22,6 +20,47 @@ from pyspark.sql.types import (
     TimestampType,
     StructField,
 )
+
+"""
+apply goldengate events log to target
+    Resolution:
+        Ingest goldengate events logs from parquet on s3
+        apply events appropriately to target table
+        commit target table as delta on s3
+
+    Logical steps:
+        0. Read In event log and get tables to be considers
+        1. Read in target table and extract schema
+        2. map event log to target schema
+        3. Get unique key list from event log
+        4. Extract Records to be considered from target
+        5. Remove Records to be considered from target
+        6. Identify first event in change log for new records
+        7. Union steps 4 and 6
+        8. Apply event log to step 7 sequentially to allow for multiple events on same record
+        9. Union applied events with unconsidered records (5 and 8)
+        10. Write to target
+
+   Noites : To run in glue_etl docker
+        1. Copy this script to /jupyter_workspace/src
+        3. execute with delta support (see readme) 
+
+
+    ToDo: 
+        refactor methods into src/lib/
+        enhance commentary
+        resolve dynamic frame read/write catalog (requires glue catalog)
+        resolve dynamic frame read write to delta tables
+
+
+"""
+__author__ = "frazer.clayton@digital.justice.gov.uk"
+
+DATABASE_NAME = "dpr-glue-catalog-development-development"
+GG_TRANSAC_EVENTS_TABLE = "offenders_orig"
+
+# use glue catalog True/False
+USE_CATALOG = False
 
 possible_types = {
     1: lambda t: DoubleType(),
@@ -53,45 +92,27 @@ def __type_for(datatype):
     return get_type(datatype)
 
 
-ddl = (
-    "OFFENDER_ID number,OFFENDER_NAME_SEQ number,ID_SOURCE_CODE varchar2(48),LAST_NAME varchar2(140),NAME_TYPE "
-    "varchar2(48),FIRST_NAME varchar2(140),MIDDLE_NAME varchar2(140),BIRTH_DATE date,SEX_CODE varchar2(48),"
-    "SUFFIX varchar2(48),LAST_NAME_SOUNDEX varchar2(24),BIRTH_PLACE varchar2(100),BIRTH_COUNTRY_CODE varchar2(48),"
-    "CREATE_DATE date,LAST_NAME_KEY varchar2(140),ALIAS_OFFENDER_ID number,FIRST_NAME_KEY varchar2(140),"
-    "MIDDLE_NAME_KEY varchar2(140),OFFENDER_ID_DISPLAY varchar2(40),ROOT_OFFENDER_ID number,CASELOAD_TYPE varchar2("
-    "48),MODIFY_USER_ID varchar2(128),MODIFY_DATETIME timestamp(9),ALIAS_NAME_TYPE varchar2(48),PARENT_OFFENDER_ID "
-    "number,UNIQUE_OBLIGATION_FLAG varchar2(4),SUSPENDED_FLAG varchar2(4),SUSPENDED_DATE date,RACE_CODE varchar2("
-    "48),REMARK_CODE varchar2(48),ADD_INFO_CODE varchar2(48),BIRTH_COUNTY varchar2(80),BIRTH_STATE varchar2(80),"
-    "MIDDLE_NAME_2 varchar2(140),TITLE varchar2(48),AGE number,CREATE_USER_ID varchar2(160),LAST_NAME_ALPHA_KEY "
-    "varchar2(4),CREATE_DATETIME timestamp(9),NAME_SEQUENCE varchar2(48),AUDIT_TIMESTAMP timestamp(9),AUDIT_USER_ID "
-    "varchar2(128),AUDIT_MODULE_NAME varchar2(260),AUDIT_CLIENT_USER_ID varchar2(256),AUDIT_CLIENT_IP_ADDRESS "
-    "varchar2(156),AUDIT_CLIENT_WORKSTATION_NAME varchar2(256),AUDIT_ADDITIONAL_INFO varchar2(1024)"
-)
+records_of_interest = {1061, 873, 141, 150, 127, 128, 129}
 
 
-def get_schema(with_event_type=False, prefix=False):
-    field_list = ddl.split(",")
-    struct_list = []
+def get_table_location(database, table_name):
+    boto_glue = boto3.client("glue")
+    table_def = boto_glue.get_table(DatabaseName=database, Name=table_name)
+    return table_def["Table"]["StorageDescriptor"]["Location"]
 
-    for field in field_list:
-        field_name, field_type = field.split(" ")
-        field_name = field_name.lower()
-        field_type = field_type.split("(")[0].lower()
-        if prefix:
-            prefix = "__"
-        else:
-            prefix = ""
-        field_name = "{}{}".format(prefix, field_name)
 
-        struct_list.append(StructField(field_name, __type_for(field_type), True))
+def format_table_name(table_name):
+    return table_name.split(".")[1].lower()
 
-    struct_list.append(StructField("{}admin_hash".format(prefix), StringType(), True))
-    struct_list.append(StructField("{}admin_gg_pos".format(prefix), StringType(), True))
-    struct_list.append(StructField("{}admin_gg_op_ts".format(prefix), TimestampType(), True))
-    struct_list.append(StructField("{}admin_event_ts".format(prefix), TimestampType(), True))
+
+def update_schema(schema, with_event_type=False, prefix=False):
+    struct_list = schema.fields
     if with_event_type:
         struct_list.append(StructField("event_type", StringType(), True))
         struct_list.append(StructField("previous_hash", StringType(), True))
+        struct_list.append(StructField("table", StringType(), True))
+        struct_list.append(StructField("schema", StringType(), True))
+
     return StructType(struct_list)
 
 
@@ -99,59 +120,12 @@ def get_primary_key():
     return "offender_id"
 
 
-"""
-apply goldengate events log to target
-
-
-"""
-__author__ = "frazer.clayton@digital.justice.gov.uk"
-
-# use glue catalog True/False
-USE_CATALOG = False
-
 temp_dataframe = None
-# configuration
-config_gg_events = dict(
-    source_bucket="dpr-demo-development-20220906101710889000000001",
-    source="data/dummy/kinesis/transac/parquet",
-    schema="oms_owner",
-    table="offenders",
-)
-config_target_table = dict(
-    target_bucket="dpr-demo-development-20220906101710889000000001",
-    target_final="data/dummy/database/",
-    schema="oms_owner",
-    table="offenders",
-    # partition_by = ["date", "time"]
-    partition_by=["create_date"],
-)
 
 
-def update_config():
-    """
-    Update configuration with elements describing paths to data
-    :return: None
-    """
-
-    config_gg_events["path"] = (
-        config_gg_events["source_bucket"]
-        + "/"
-        + config_gg_events["source"]
-        + "/"
-        + config_gg_events["schema"]
-        + "/"
-        + config_gg_events["table"]
-    )
-
-    config_target_table["path"] = (
-        config_target_table["target_bucket"]
-        + "/"
-        + config_target_table["target_final"]
-        + "/"
-        + config_target_table["schema"]
-        + "/"
-        + config_target_table["table"]
-    )
+def get_target_table_name(gg_table_name):
+    _tt_name = gg_table_name.split(".")[1].lower()
+    return _tt_name
 
 
 def write_catalog(gluecontext, config, frame):
@@ -177,12 +151,15 @@ def write_catalog(gluecontext, config, frame):
 
 def write_s3(gluecontext, config, frame):
     """
-    write output to S3
+    write glue dynamic frame to S3
+        CURRENTLY THIS IS NOT WORKING WITH DELTA FORMAT
     :param gluecontext: Glue context
     :param config: configuration dictionary
     :param frame: Glue Dynamic Frame
     :return:None
     """
+
+    print("###### s3 writing to {}".format(config["path"]))
     gluecontext.write_dynamic_frame.from_options(
         frame=frame,
         connection_type="s3",
@@ -194,23 +171,48 @@ def write_s3(gluecontext, config, frame):
     )
 
 
-def write_frame(gluecontext, config, frame):
+def read_table_to_df(gluecontext, database, table_name, file_format="parquet"):
     """
-    wrapper for write mechanism determined by USE_CATALOG
+    Read from S3 into Dataframe
     :param gluecontext: Glue context
     :param config: configuration dictionary
-    :param frame: Glue Dynamic Frame
-    :return:None
+    :param key_suffix: suffix to be added to path
+    :return: spark dataframe
     """
-    if USE_CATALOG:
-        write_catalog(gluecontext=gluecontext, config=config, frame=frame)
-    else:
-        write_s3(gluecontext=gluecontext, config=config, frame=frame)
+    read_path = get_table_location(database, table_name)
+
+    input_dydf = gluecontext.create_dynamic_frame_from_options(
+        connection_type="s3",
+        connection_options={"path": read_path},
+        format=file_format,
+    )
+    return input_dydf.toDF()
+
+
+def read_delta_table(database, table_name):
+    # Write data as DELTA TABLE
+
+    frame = spark.read.format("delta").load(get_table_location(database, table_name))
+    return frame
+
+    # Generate MANIFEST file for Athena/Catalog
+    # deltaTable = DeltaTable.forPath(spark, "s3://{}/".format(config["path_delta"]))
+    # deltaTable.generate("symlink_format_manifest")
+
+
+def write_delta_table(database, table_name, frame):
+    # Write data as DELTA TABLE
+    frame.write.format("delta").mode("overwrite").save(get_table_location(database, table_name))
+
+    # Generate MANIFEST file for Athena/Catalog
+    # deltaTable = DeltaTable.forPath(spark, "s3://{}/".format(config["path"]))
+    # deltaTable.generate("symlink_format_manifest")
 
 
 def read_s3_to_df(gluecontext, config, key_suffix=None, file_format="parquet"):
     """
     Read from S3 into Dataframe
+    :param file_format: file format
     :param gluecontext: Glue context
     :param config: configuration dictionary
     :param key_suffix: suffix to be added to path
@@ -231,20 +233,24 @@ def create_empty_df(schema):
     return spark.createDataFrame(data=spark.sparkContext.emptyRDD(), schema=schema)
 
 
-def load_sample_to_df(df):
-    dfsample = df.sample(0.01)
+def load_sample_to_df(df, sample=0.01):
+    """
+    sample df records
+    :param sample: sample decimal
+    :param df: dataframe to sample
+    :return: sampled dataframe
+    """
+    dfsample = df.sample(sample)
     print(dfsample.count())
     return dfsample
 
 
-def output_df():
-    target_schema = get_schema()
-
-    temp_df = create_empty_df(target_schema)
-    return temp_df
-
-
 def get_schema_fields_as_dict(schema):
+    """
+    convert schem to dictionary
+    :param schema: schema
+    :return: dictionary containing schema
+    """
     schema_dict = {}
     for fld in schema:
         schema_dict[fld.name] = None
@@ -252,13 +258,26 @@ def get_schema_fields_as_dict(schema):
 
 
 def get_schema_field_type(schema, fldname):
-    schema_dict = {}
+    """
+    get datatype of field name from schema
+    :param schema: schema
+    :param fldname: field name
+    :return: datatype
+    """
+
     for fld in schema:
         if fld.name == fldname:
             return fld.dataType
 
 
 def format_field(schema, fldname, fld_val):
+    """
+    format field value based on schema datatypes
+    :param schema: schema
+    :param fldname: fieldname
+    :param fld_val: field value
+    :return: field value as datatype
+    """
     fldtype = get_schema_field_type(schema=schema, fldname=fldname)
     new_val = fld_val
     if fld_val is not None:
@@ -272,10 +291,12 @@ def format_field(schema, fldname, fld_val):
 
 
 def mapper(row_in, schema):
-    # print("row_in")
-    # print(row_in["after"])
-    # new_row = Row()
-    # print("#####")
+    """
+    map rows to schema
+    :param row_in: row containing unmapped record
+    :param schema: schema definition
+    :return: row type in correct schema format
+    """
 
     new_row_dict = get_schema_fields_as_dict(schema)
 
@@ -287,9 +308,7 @@ def mapper(row_in, schema):
     row_dict = row_out.asDict()
 
     for fld_name in row_dict:
-        if fld_name.lower() == "modified_datetime":
-            new_row_dict["modify_datetime"] = row_dict[fld_name]
-        else:
+        if fld_name.lower() in new_row_dict:
             new_row_dict[fld_name.lower()] = format_field(
                 schema=schema, fldname=fld_name.lower(), fld_val=row_dict[fld_name]
             )
@@ -300,16 +319,30 @@ def mapper(row_in, schema):
     new_row_dict["admin_gg_op_ts"] = format_field(schema=schema, fldname="admin_gg_op_ts", fld_val=row_in["op_ts"])
     new_row_dict["admin_event_ts"] = datetime.datetime.now()
     new_row_dict["event_type"] = row_in["op_type"]
+    new_row_dict["table"] = row_in["table"].split(".")[1]
+    new_row_dict["schema"] = row_in["table"].split(".")[0]
     # print(new_row_dict)
     return Row(**new_row_dict)
 
 
 def convert_to_dict_list(frame):
+    """
+    convert all rows to dictionary list
+    :param frame: dataframe
+    :return: list of dictionarys containing row data
+    """
     out_dict = frame.rdd.map(lambda row: row.asDict()).collect()
     return out_dict
 
 
 def compare_dicts(dict_orig, row_in, key_field):
+    """
+    compare the two dictionary records and update original based on event type
+    :param dict_orig: original dict
+    :param row_in: event record
+    :param key_field: record key
+    :return: updated original dict
+    """
     if row_in["previous_hash"] == dict_orig["admin_hash"]:
         if row_in["event_type"] == "U":
             for key in dict_orig:
@@ -332,18 +365,34 @@ def compare_dicts(dict_orig, row_in, key_field):
 
 
 def apply_events(row_in, key_field, event_dict):
+    """
+    apply the events to the row sequentially
+    :param row_in: original row
+    :param key_field: record key
+    :param event_dict: list of events as dict
+    :return: updated row
+    """
     row_dict = row_in.asDict()
 
     tmp_dict = []
     for dct in event_dict:
         if dct[key_field] == row_dict[key_field]:
             tmp_dict.append(dct)
-    print("tmp dict len {}".format(len(tmp_dict)))
 
     for event_rec in tmp_dict:
         row_dict = compare_dicts(row_dict, event_rec, key_field)
 
     return Row(**row_dict)
+
+
+def get_distinct_column_values_from_df(frame, column):
+    out_list = []
+
+    row_array = frame.select(column).distinct().collect()
+    for _row in row_array:
+        out_list.append(_row[column])
+
+    return out_list
 
 
 def rename_columns(frame):
@@ -358,12 +407,14 @@ def show_table(table_df):
     table_df.select(
         col("offender_id"),
         col("title"),
+        col("first_name"),
+        col("last_name"),
         col("create_date"),
         col("admin_hash"),
         col("admin_gg_pos"),
         col("admin_event_ts"),
         # col("__action"),
-    ).filter((col("offender_id").isin({1061, 873, 127, 128, 129}))).show(10)
+    ).filter((col("offender_id").isin(records_of_interest))).show(10, truncate=False)
 
     print("##########################################")
 
@@ -372,15 +423,17 @@ def show_events(event_df):
     print("##########################################")
     print("event records:", event_df.count())
     event_df.select(
+        col("table"),
         col("offender_id"),
-        col("title"),
         col("create_date"),
         col("admin_hash"),
         col("previous_hash"),
         col("admin_gg_pos"),
         col("admin_event_ts"),
         col("event_type"),
-    ).filter((col("offender_id").isin({1061, 873, 127, 128, 129}))).sort("offender_id", "admin_gg_pos").show(30)
+    ).filter((col("offender_id").isin(records_of_interest))).sort("offender_id", "admin_gg_pos").show(
+        30, truncate=False
+    )
     print("##########################################")
 
 
@@ -391,24 +444,28 @@ def show_bef_after_applied(df_to_consider, df_applied):
     df_to_consider.select(
         col("offender_id"),
         col("title"),
+        col("first_name"),
+        col("last_name"),
         col("create_date"),
         col("admin_hash"),
         col("admin_gg_pos"),
         col("admin_event_ts"),
         col("__action"),
-    ).filter((col("offender_id").isin({1061, 873, 127, 128, 129}))).show(10)
+    ).filter((col("offender_id").isin(records_of_interest))).show(10)
 
     print("records applied:", df_applied.count())
     print("example")
     df_applied.select(
         col("offender_id"),
         col("title"),
+        col("first_name"),
+        col("last_name"),
         col("create_date"),
         col("admin_hash"),
         col("admin_gg_pos"),
         col("admin_event_ts"),
         col("__action"),
-    ).filter((col("offender_id").isin({1061, 873, 127, 128, 129}))).show(50)
+    ).filter((col("offender_id").isin(records_of_interest))).show(50)
 
     print("##########################################")
 
@@ -423,90 +480,27 @@ def start():
 
     glueContext = GlueContext(SparkContext.getOrCreate())
 
-    update_config()
+    """0. Read In event log and get tables to be considers"""
 
-    temp_schema = get_schema(with_event_type=True)
-    target_schema = get_schema()
-    target_key = get_primary_key()
+    target_table = "offender_booking"
+    target_table_orig = target_table + "_orig"
+    print(target_table)
+    df_table_in = read_delta_table(database=DATABASE_NAME, table_name=GG_TRANSAC_EVENTS_TABLE)
 
-    df_event_log = read_s3_to_df(gluecontext=glueContext, config=config_gg_events)  # , key_suffix="date=2022-09-13")
-
-    df_table_in = read_s3_to_df(gluecontext=glueContext, config=config_target_table)
-    # df_table_in = df_table_in.withColumn("status", lit(0))
-
-    # df_event_log.show()
-    # df_table_in.show()
-
-    show_table(df_table_in)
-
-    df_event_log = df_event_log.rdd.map(lambda row: mapper(row_in=row, schema=temp_schema)).toDF(schema=temp_schema)
-
-    df_unique_key = rename_columns(frame=df_event_log.select(target_key).distinct())
-
-    # consider events against existing records
-
-    show_events(df_event_log)
-
-    df_to_consider = df_table_in.join(
-        df_unique_key, df_table_in[target_key] == df_unique_key["__{}".format(target_key)], "inner"
-    ).drop("__{}".format(target_key))
-
-    df_to_consider = df_to_consider.withColumn("__action", lit(""))
-    action_schema = df_to_consider.schema
-
-    df_event_log = df_event_log.sort("admin_gg_pos")
-
-    temp_dict_list = convert_to_dict_list(df_event_log)
-
-    df_applied = df_to_consider.rdd.map(
-        lambda row: apply_events(row_in=row, key_field=target_key, event_dict=temp_dict_list)
-    ).toDF(schema=action_schema)
-
-    show_bef_after_applied(df_to_consider, df_applied)
-
-    # consider events against non existing records
-    ##########################################
-
-    # get primary events where not in existing considered
-
-    # df_unique_key.filter(col("__offender_id").isin({1061,873,127, 128, 129})).show(10)
-    df_unique_applied_key = frame = df_applied.select(target_key).distinct()
-
-    # df_unique_applied_key.filter(col("offender_id").isin({1061,873,127, 128, 129})).show(10)
-
-    df_unique_key = (
-        df_unique_key.join(
-            df_unique_applied_key, df_unique_applied_key[target_key] == df_unique_key["__{}".format(target_key)], "left"
-        )
-        .filter(col(target_key).isNull())
-        .drop(target_key)
-    )
-
-    # df_unique_key.filter(col("__offender_id").isin({1061,873,127, 128, 129})).show(10)
-    # df_unique_key = df_unique_key.join(df_applied,df_unique_key[target_key] ==  df_unique_key["__{}".format(target_key)],"right") #.filter(col(target_key).isNull())
-
-    w = Window.partitionBy(target_key)
-    df_primary_events = (
-        df_event_log.withColumn("minpos", min("admin_gg_pos").over(w))
-        .where(col("admin_gg_pos") == col("minpos"))
-        .drop("minpos")
-    )
-
-    df_primary_events = df_primary_events.drop("event_type").drop("previous_hash")
-
-    df_to_consider = df_primary_events.join(
-        df_unique_key, df_primary_events[target_key] == df_unique_key["__{}".format(target_key)], "inner"
-    ).drop("__{}".format(target_key))
-
-    df_to_consider = df_to_consider.withColumn("__action", lit(""))
-    action_schema = df_to_consider.schema
-
-    df_added = df_to_consider.rdd.map(
-        lambda row: apply_events(row_in=row, key_field=target_key, event_dict=temp_dict_list)
-    ).toDF(schema=action_schema)
-
-    show_bef_after_applied(df_to_consider, df_added)
+    df_table_in.select(
+        col("offender_id"),
+        col("title"),
+        col("first_name"),
+        col("last_name"),
+        col("create_date"),
+        col("admin_hash"),
+        col("admin_gg_pos"),
+        col("admin_event_ts"),
+    ).sort("offender_id").show(20)
+    # .filter(col("offender_id").isin(127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143,144,145))
 
 
 if __name__ == "__main__":
+    from pyspark.shell import spark
+
     start()
